pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "article-checklist-manager"
version = "0.0.0"
authors = [{name="ACM Dev"}]
requires-python = ">=3.8"
dependencies = [
    "typer>=0.7",
    "PyYAML>=6.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4",
    "pytest-cov>=4.0"
]

[tool.pytest.ini_options]
addopts = "-vv --cov=acm"
